BetWriter/requirements.txt
sqlalchemy==1.4.44
psycopg2-binary
kafka-python==2.0.2